form_test1/my_project/pyproject.toml
[project]
name = "lifeblood_equipment_check"
version = "0.0.1"
authors = [{ name = "olivia.ren@databricks.com" }]
requires-python = ">= 3.11"
dependencies = [
    "streamlit>=1.28.0",
    "pandas>=2.0.0",
    "databricks-sql-connector>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest",

    # Code completion support for Lakeflow Declarative Pipelines, also install databricks-connect
    "databricks-dlt",

    # databricks-connect can be used to run parts of this project locally.
    # See https://docs.databricks.com/dev-tools/databricks-connect.html.
    #
    # Note, databricks-connect is automatically installed if you're using Databricks
    # extension for Visual Studio Code
    # (https://docs.databricks.com/dev-tools/vscode-ext/dev-tasks/databricks-connect.html).
    #
    # To manually install databricks-connect, uncomment the line below to install a version
    # of db-connect that corresponds to the Databricks Runtime version used for this project.
    # See https://docs.databricks.com/dev-tools/databricks-connect.html
    # "databricks-connect>=15.4,<15.5",
]

[tool.pytest.ini_options]
pythonpath = "src"
testpaths = [
    "tests",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/my_project"]

[project.scripts]
main = "my_project.main:main"
